micromegas_telemetry_sink::stream_block

Trait StreamBlock

pub trait StreamBlock: Send + Sync {
    // Required method
    fn encode_bin(&self, process_info: &ProcessInfo) -> Result<Vec<u8>>;
}

Required Methods

fn encode_bin(&self, process_info: &ProcessInfo) -> Result<Vec<u8>>

Encodes the stream block into a binary format.

This function serializes the block data, compresses it, and then encodes it into the wire format for transmission.

Arguments

process_info - Information about the current process, used for time calibration.